updates the setup  script to only allow all and develop once

+++ setup_env.sh
@@ -21,6 +21,11 @@ DEVNET=false
 BOTS=false
 PORTS=false
 
+## New variables
+ALL_FLAG=false
+DEVELOP_FLAG=false
+FLAGS_COUNT=0
+
 ## Loop through the arguments
 while [[ $# -gt 0 ]]; do
     case $1 in
@@ -28,20 +33,27 @@ while [[ $# -gt 0 ]]; do
             DEVNET=true
             BOTS=true
             PORTS=true
+            ALL_FLAG=true
+            FLAGS_COUNT=$((FLAGS_COUNT+1))
             ;;
         --develop)
             DEVNET=true
             BOTS=false
             PORTS=true
+            DEVELOP_FLAG=true
+            FLAGS_COUNT=$((FLAGS_COUNT+1))
             ;;
         --devnet)
             DEVNET=true
+            FLAGS_COUNT=$((FLAGS_COUNT+1))
             ;;
         --bots)
             BOTS=true
+            FLAGS_COUNT=$((FLAGS_COUNT+1))
             ;;
         --ports)
             PORTS=true
+            FLAGS_COUNT=$((FLAGS_COUNT+1))
             ;;
         *)
             echo "Unknown flag: $1"
@@ -49,6 +61,12 @@ while [[ $# -gt 0 ]]; do
     esac
     shift
 done
+
+# Throw an error if `--all` or `--develop` is not specified by itself
+if { $ALL_FLAG || $DEVELOP_FLAG; } && [ $FLAGS_COUNT -gt 1 ]; then
+    echo "Error: --all or --develop must be specified by itself."
+    exit 1
+fi
 
 # Make a new .env file
 rm -f .env
